+++ backend/src/main/java/wooteco/prolog/roadmap/application/dto/KeywordResponse.java
@@ -5,7 +5,7 @@ import java.util.Set;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import wooteco.prolog.roadmap.Keyword;
+import wooteco.prolog.roadmap.domain.Keyword;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
