+++ backend/src/main/java/wooteco/prolog/roadmap/domain/Keyword.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.roadmap;
+package wooteco.prolog.roadmap.domain;
 
 import java.util.HashSet;
 import java.util.Set;
