+++ backend/src/main/java/wooteco/prolog/roadmap/domain/repository/QuizRepository.java
@@ -1,7 +1,7 @@
-package wooteco.prolog.roadmap.repository;
+package wooteco.prolog.roadmap.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.prolog.roadmap.Keyword;
+import wooteco.prolog.roadmap.domain.Keyword;
 
 public interface QuizRepository extends JpaRepository<Keyword, Long> {
